remove vector convertion from SearchIVF function

+++ pkg/utils/getLimit.go
@@ -1,12 +1,12 @@
 package utils
 
-func GetLimit(value float64, maxValue float64) float64 {
-	division := float64(value) / float64(maxValue)
+func GetLimit(value float64, maxValue float64) float32 {
+	division := value / maxValue
 	if division > 1 {
 		return 1.0
 	}
 	if division < 0 {
 		return 0
 	}
-	return float64(division)
+	return float32(division)
 }
